cherrypy starter kit with postgres/requirements.txt
CherryPy-SQLAlchemy
cherrypy-cors
CherryPy
psycopg2-binary

